Renamed files to be consistent with new class names Sample and SamplePoint
