+++ CMakeLists.txt
@@ -135,37 +135,44 @@ include_directories(${CMAKE_CURRENT_SOURCE_DIR}/thirdparty/Catch)
 # These are the sources we need for compilation of the library
 set(SRC_LIST
     include/approximant.h
+    include/saveable.h
     include/bsplineapproximant.h
     include/bspline.h
     include/bsplinebasis.h
     include/bsplinebasis1d.h
     include/cinterface.h
-    include/datasample.h
-    include/datatable.h
+    include/samplepoint.h
+    include/sample.h
     include/function.h
+    include/linearfunction.h
     include/definitions.h
     include/linearsolvers.h
     include/mykroneckerproduct.h
+    include/polynomial.h
     include/polynomialapproximant.h
     include/psplineapproximant.h
     include/rbfapproximant.h
     include/rbfterm.h
     include/serializer.h
     include/utilities.h
+    include/ols.h
     src/bsplineapproximant.cpp
     src/bspline.cpp
     src/bsplinebasis.cpp
     src/bsplinebasis1d.cpp
     src/cinterface.cpp
-    src/datasample.cpp
-    src/datatable.cpp
+    src/samplepoint.cpp
+    src/sample.cpp
     src/function.cpp
+    src/linearfunction.cpp
     src/mykroneckerproduct.cpp
+    src/polynomial.cpp
     src/polynomialapproximant.cpp
     src/psplineapproximant.cpp
     src/rbfapproximant.cpp
     src/serializer.cpp
     src/utilities.cpp
+    src/ols.cpp
 )
 set(TEST_SRC_LIST
     ${SRC_LIST}
@@ -174,9 +181,10 @@ set(TEST_SRC_LIST
     test/approximation/psplineapproximant.cpp
     test/approximation/rbfapproximant.cpp
     test/approximation/polynomialapproximant.cpp
-    test/general/datatable.cpp
+    test/general/bsplineapproximant.cpp
+    test/general/sample.cpp
     test/general/utilities.cpp
-    test/serialization/datatable.cpp
+    test/serialization/sample.cpp
     test/serialization/bsplineapproximant.cpp
     test/serialization/psplineapproximant.cpp
     test/serialization/rbfapproximant.cpp
