+++ src/samplepoint.cpp
@@ -7,7 +7,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */
 
-#include "datasample.h"
+#include "samplepoint.h"
 
 namespace SPLINTER
 {
