+++ test/serialization/polynomialapproximant.cpp
@@ -8,7 +8,7 @@
 */
 
 #include <Catch.h>
-#include <datatable.h>
+#include <sample.h>
 #include "polynomialapproximant.h"
 #include "testingutilities.h"
